Try bincode rather than zstd
