+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
   "core",
   "backend",
+  "backend/backend-types",
   "frontend/shengji-wasm"
 ]
