+++ backend/Cargo.toml
@@ -12,8 +12,9 @@ dynamic = ["slog-term"]
 
 [dependencies]
 anyhow = "1.0"
+shengji-types = { path = "./backend-types" }
+bincode = "1.3"
 futures = { version = "0.3" }
-static_dir = "0.1.3"
 lazy_static = "1.4.0"
 phf = "0.8.0"
 serde = { version = "1.0", features = ["derive"] }
@@ -23,9 +24,9 @@ slog = "2.5"
 slog-async = "2.5"
 slog-bunyan = "2.2"
 slog-term = { version = "2.5", optional = true }
+static_dir = "0.1.3"
 tokio = { version = "0.2", features = ["macros", "fs"] }
 warp = "0.2"
-zstd = "0.5"
 
 [build-dependencies]
 vergen = "3"
