+++ backend/backend-types/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "shengji-types"
+version = "0.1.0"
+authors = ["Robert Ying <rbtying@aeturnalus.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+shengji-core = { path = "../../core" }
