+++ backend/backend-types/src/lib.rs
@@ -1,0 +1,21 @@
+use serde::{Deserialize, Serialize};
+use shengji_core::{game_state, interactive};
+
+#[allow(clippy::large_enum_variant)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum GameMessage {
+    State {
+        state: game_state::GameState,
+    },
+    Message {
+        from: String,
+        message: String,
+    },
+    Broadcast {
+        data: interactive::BroadcastMessage,
+        message: String,
+    },
+    Beep,
+    Error(String),
+    Kicked,
+}
