+++ frontend/shengji-wasm/Cargo.toml
@@ -15,11 +15,13 @@ crate-type = ["cdylib"]
 lto = true
 
 [dependencies]
+bincode = "1.3"
 console_error_panic_hook = "0.1.5"
 serde = { version = "1.0", features = ["derive"]}
+serde_json = "1.0"
 shengji-core = { path = "../../core" }
+shengji-types = { path = "../../backend/backend-types" }
 smallvec = { version = "1.3.0", features = ["serde"] }
 wasm-bindgen = { version = "0.2.45", features = ["serde-serialize"] }
 web-sys = { version = "0.3.22", features = ["console"]}
 wee_alloc = { version = "0.4.2"}
-ruzstd = "0.2.1"
