+++ frontend/src/WasmProvider.tsx
@@ -32,7 +32,7 @@ const ShengjiProvider = (props: IProps): JSX.Element => {
         computeScore: (req) => {
           return Shengji.compute_score(req);
         decodeWireFormat: (req) => {
-          return JSON.parse(Shengji.zstd_decompress(req));
+          return JSON.parse(Shengji.bincode_deserialize(req));
         },
       }}
     >
